API: clearly define public/private APIs
